+++ app/product/biz/service/list_products.go
@@ -39,7 +39,7 @@ func (s *ListProductsService) Run(req *product.ListProductsReq) (resp *product.L
 	resp = &product.ListProductsResp{}
 	for _, v1 := range c {
 		for _, v := range v1.Products {
-			resp.Products = append(resp.Products, &product.Product{Id: uint32(v.ID), Name: v.Name, Description: v.Description, Picture: v.Picture, Price: v.Price})
+			resp.Products = append(resp.Products, &product.Product{Id: v.ID, Name: v.Name, Description: v.Description, Picture: v.Picture, Price: v.Price})
 		}
 	}
